Add files via upload
This course provided comprehensive training on the methodologies and tools for effective online learning and teaching. It covered various aspects including instructional design, online pedagogy, and the use of technology to enhance the learning experience. Through this course, I gained valuable insights and practical skills that are essential for creating and managing successful online educational programs.